working lower bound

+++ HW5/src/task/TspJob.java
@@ -52,7 +52,7 @@ public class TspJob implements Job {
         List<Integer> partialTrip = new ArrayList<>();
         partialTrip.add(0);
 
-        TaskTsp startTask = new TaskTsp(new Closure(0, new Global(new Long(Long.MAX_VALUE)), new Continuation("-1",-1,new TaskTsp.Wrapper(unusedCities, partialTrip))));
+        TaskTsp startTask = new TaskTsp(new Closure(0, new Global(new Double(Double.MAX_VALUE)), new Continuation("-1",-1,new TaskTsp.Wrapper(unusedCities, partialTrip))));
 
 /*      //Used for Simmulated Annealing
         Integer[] startPath = new Integer[cities.length];
